requirements.txt
flask
flask_restx
requests
playwright
instagrapi
pillow
pyyaml
peewee
psycopg2-binary
pycryptodome